[HDJUNCTION-8] 셀프리뷰 및 코드리팩토링
1] Service interface 제거
    (서비스를 분리하고자 할 때 interface는 불필요하여도 강제구현이 필요할 수 있으므로)
2] Patient(환자) Update -> dirty checking사용으로 save메소드 불필요 제거
3] package명 rename (entity -> domain)
4] Code, CodeGroup 로직에서 사용하지않으므로 불필요 어노테이션 정리
- Cascade 제거
- DynamicInsert, DynamicUpdate 제거
5] @ManyToOne LAZY로딩 타입 부여
6] printstacktrace() 제거
7] ConstantUtil 인스턴스 만들지 못하도록 처리
8] PatientService, VisitService method명 rename
9] 환자 목록 조회 API 리팩토링
10] Gender enum -> name대신 enum 리턴하도록 개선
    (클라이언트에서 어떤 필드를 사용해야할지 직접 고민하도록 설계)
