+++ src/main/java/com/hdjunction/project/yunsang/domain/hospital/domain/Code.java
@@ -1,20 +1,17 @@
-package com.hdjunction.project.yunsang.domain.hospital.entity;
+package com.hdjunction.project.yunsang.domain.hospital.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.hibernate.annotations.DynamicInsert;
-import org.hibernate.annotations.DynamicUpdate;
 
 @Entity
 @Table(name = "code")
-@DynamicInsert
-@DynamicUpdate
 @Getter
 @NoArgsConstructor
 public class Code {
@@ -23,7 +20,7 @@ public class Code {
     private String code;
     @Column(name = "code_name")
     private String codeName;
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "code_group")
     private CodeGroup codeGroup;
 }
