+++ src/main/java/com/hdjunction/project/yunsang/domain/hospital/domain/CodeGroup.java
@@ -1,6 +1,5 @@
-package com.hdjunction.project.yunsang.domain.hospital.entity;
+package com.hdjunction.project.yunsang.domain.hospital.domain;
 
-import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
@@ -8,16 +7,12 @@ import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.hibernate.annotations.DynamicInsert;
-import org.hibernate.annotations.DynamicUpdate;
 
 import java.util.ArrayList;
 import java.util.List;
 
 @Entity
 @Table(name = "code_group")
-@DynamicInsert
-@DynamicUpdate
 @Getter
 @NoArgsConstructor
 public class CodeGroup {
@@ -28,6 +23,6 @@ public class CodeGroup {
     private String codeGroupName;
     @Column(name = "description")
     private String description;
-    @OneToMany(mappedBy = "codeGroup", orphanRemoval = true, cascade = CascadeType.ALL)
+    @OneToMany(mappedBy = "codeGroup")
     private List<Code> code = new ArrayList<>();
 }
