+++ src/main/java/com/hdjunction/project/yunsang/domain/hospital/domain/Patient.java
@@ -1,9 +1,10 @@
-package com.hdjunction.project.yunsang.domain.hospital.entity;
+package com.hdjunction.project.yunsang.domain.hospital.domain;
 
 import com.hdjunction.project.yunsang.domain.hospital.dto.PatientRequestDto;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -43,7 +44,7 @@ public class Patient {
     private String birth;
     @Column(name = "phone")
     private String phone;
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "hospital_id", insertable = false, updatable = false)
     private Hospital hospital;
     @OneToMany(mappedBy = "patient"
