+++ src/main/java/com/hdjunction/project/yunsang/domain/hospital/repository/HospitalRepository.java
@@ -1,6 +1,6 @@
 package com.hdjunction.project.yunsang.domain.hospital.repository;
 
-import com.hdjunction.project.yunsang.domain.hospital.entity.Hospital;
+import com.hdjunction.project.yunsang.domain.hospital.domain.Hospital;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface HospitalRepository extends JpaRepository<Hospital, Long> {
