+++ src/main/java/com/hdjunction/project/yunsang/domain/hospital/repository/PatientRepository.java
@@ -1,6 +1,6 @@
 package com.hdjunction.project.yunsang.domain.hospital.repository;
 
-import com.hdjunction.project.yunsang.domain.hospital.entity.Patient;
+import com.hdjunction.project.yunsang.domain.hospital.domain.Patient;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PatientRepository extends JpaRepository<Patient, Long>, PatientRepositoryCustom {
