+++ src/main/java/com/hdjunction/project/yunsang/domain/hospital/repository/VisitRepository.java
@@ -1,6 +1,6 @@
 package com.hdjunction.project.yunsang.domain.hospital.repository;
 
-import com.hdjunction.project.yunsang.domain.hospital.entity.Visit;
+import com.hdjunction.project.yunsang.domain.hospital.domain.Visit;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface VisitRepository extends JpaRepository<Visit, Long> {
